Convert a bit more complex set of helpers to Ts
