+++ e2e/support/helpers/e2e-notebook-helpers.ts
@@ -1,0 +1,197 @@
+import { popover } from "e2e/support/helpers/e2e-ui-elements-helpers";
+
+/**
+ *
+ * @TODO: Improve description and chaining
+ */
+export function openNotebook() {
+  return cy.icon("notebook").click();
+}
+
+/**
+ * Select a specific notebook step like filter, join, breakout, etc.
+ *
+ * @see {@link https://github.com/metabase/metabase/pull/17708#discussion_r700082403}
+ */
+export function getNotebookStep(
+  type:
+    | "summarize"
+    | "join"
+    | "expression"
+    | "filter"
+    | "sort"
+    | "limit"
+    | "data",
+  { stage = 0, index = 0 } = {},
+): Cypress.Chainable<JQuery<HTMLElement>> {
+  return cy.findByTestId(`step-${type}-${stage}-${index}`);
+}
+
+/**
+ * @summary Visualize notebook query results.
+ *
+ * This helper intelligently waits for the query to load, and gives you an option
+ * to assert on the waited query response.
+ *
+ * @example
+ * visualize();
+ *
+ * visualize(response => {
+ *   expect(response.body.error).to.not.exist;
+ * });
+ */
+export function visualize(callback?: (response?: object) => void) {
+  cy.intercept("POST", "/api/dataset").as("dataset");
+
+  cy.button("Visualize").click();
+
+  cy.wait("@dataset").then(({ response }) => {
+    if (callback) {
+      callback(response);
+    }
+  });
+}
+
+/**
+ * Summarize (Aggregate).
+ *
+ * Doesn't support summarizing using Custom Expression!
+ */
+export function addSummaryField({
+  metric,
+  table,
+  field,
+  stage = 0,
+  index = 0,
+}: {
+  metric:
+    | "Count of rows"
+    | "Sum of ..."
+    | "Average of ..."
+    | "Number of distinct values of ..."
+    | "Cumulative sum of ..."
+    | "Cumulative count of rows"
+    | "Standard deviation of ..."
+    | "Minimum of ..."
+    | "Maximum of ...";
+  table?: string;
+  field?: string;
+  stage?: number;
+  index?: number;
+}) {
+  getNotebookStep("summarize", { stage, index })
+    .findByTestId("aggregate-step")
+    .findAllByTestId("notebook-cell-item")
+    .last()
+    .click();
+
+  popover().within(() => {
+    cy.findByText(metric).click();
+    if (table) {
+      cy.findByText(table).click();
+    }
+    if (field) {
+      cy.findByText(field).click();
+    }
+  });
+}
+
+/**
+ * Group by.
+ */
+export function addSummaryGroupingField({
+  table,
+  field,
+  stage = 0,
+  index = 0,
+}: {
+  table?: string;
+  field: string;
+  stage?: number;
+  index?: number;
+}) {
+  getNotebookStep("summarize", { stage, index })
+    .findByTestId("breakout-step")
+    .findAllByTestId("notebook-cell-item")
+    .last()
+    .click();
+
+  popover().within(() => {
+    if (table) {
+      cy.findByText(table).click();
+    }
+    cy.findByText(field).click();
+  });
+}
+
+/**
+ * Remove Group by.
+ */
+export function removeSummaryGroupingField({
+  field,
+  stage = 0,
+  index = 0,
+}: {
+  field: string;
+  stage: number;
+  index: number;
+}) {
+  getNotebookStep("summarize", { stage, index })
+    .findByTestId("breakout-step")
+    .findByText(field)
+    .icon("close")
+    .click();
+}
+
+/**
+ * Join a raw table given a table and optional LHS and RHS column names
+ * (for cases when join condition can't be selected automatically).
+ *
+ * Expects a join popover to be open!
+ *
+ */
+export function joinTable(
+  tableName: string,
+  lhsColumnName?: string,
+  rhsColumnName?: string,
+) {
+  popover().findByText(tableName).click();
+  if (lhsColumnName && rhsColumnName) {
+    popover().findByText(lhsColumnName).click();
+    popover().findByText(rhsColumnName).click();
+  }
+}
+
+/**
+ * Open a saved question and join it with another saved question.
+ *
+ * Depends on a `startNewQuestion()` helper to work properly!
+ *
+ * @todo Either decouple this dependency or use `startNewQuestion()` directly here.
+ *
+ * @example
+ * startNewQuestion();
+ * selectSavedQuestionsToJoin("Q1", "Q2");
+ */
+export function selectSavedQuestionsToJoin(
+  firstQuestionName: string,
+  secondQuestionName: string,
+) {
+  cy.intercept("GET", "/api/database/*/schemas").as("loadSchemas");
+  // eslint-disable-next-line no-unscoped-text-selectors -- deprecated usage
+  cy.findByText("Saved Questions").click();
+  // eslint-disable-next-line no-unscoped-text-selectors -- deprecated usage
+  cy.findByText(firstQuestionName).click();
+  cy.wait("@loadSchemas");
+
+  // join to question b
+  cy.icon("join_left_outer").click();
+
+  popover().within(() => {
+    cy.findByText("Sample Database").should("be.visible").click();
+    cy.findByText("Raw Data").should("be.visible").click();
+    cy.findByText("Saved Questions").should("be.visible").click();
+    cy.findByText(secondQuestionName).click();
+    cy.button("Save").click();
+  });
+}
